DemandMonitorTable: types refactor / sortHardwareDemandPerformance: params and functionality refactored
